오류 수정중

+++ editor/src/CDK.Drawing/String/StringDisplay.cs
@@ -221,7 +221,7 @@ namespace CDK.Drawing
                     {
                         Type = StringParagraphType.Stroke,
                         Color0 = color,
-                        StrokeWidth = width
+                        StrokeWidth = (byte)width
                     };
                     paragraphs.Add(p);
                     return true;
